refacto run only nbAliveParticles thread of simu

+++ DemoParticles/Src/Content/ShaderStructures.h
@@ -136,7 +136,7 @@ namespace DemoParticles
 
     struct InitIndirectComputeArgs1DConstantBuffer
     {
-        UINT nbThreadGroupX;
+        float nbThreadGroupX;
 
         UINT padding[3];
     };
